Optimize analysis pipeline vectorization and fix Grayscale8 PSNR calculation (v1.0.42)
- Optimize AnalysisEngine::ssim with single-pass integer 8x8 covariance accumulation
- Vectorize AnalysisEngine::psnr sumSq and fix Grayscale8 MSE division formula
- Accelerate core analyzers (Brightness, Contrast, ColorCast, Exposure, Entropy, DeadPixel, Blur, Sharpness, Noise) with format hoisting and integer math
- Specialize AnalysisEngine::computeStatsROI Grayscale8 histogram generation and eliminate per-pixel clamps
- Fix subregion copy 64-bit index arithmetic in PSNRAnalyzer and SSIMAnalyzer
- Refactor startReportExport and computeStatsROI helpers to maintain zero complexity gate hard violations
- Bump version to 1.0.42 across CMakeLists.txt, CHANGELOG.md, and STATUS.md

+++ src/core/analyzer/SSIMAnalyzer.cpp
@@ -32,14 +32,16 @@ bool SSIMAnalyzer::analyzeRegion(const ImageFrame &frame, const mviewer::domain:
     const int cppT = vT.channelsPerPixel();
     for (int y = 0; y < rh; ++y)
     {
-        std::memcpy(subR.buffer->data() + static_cast<size_t>(y) * rw * cppR,
+        std::memcpy(subR.buffer->data() +
+                        static_cast<size_t>(y) * static_cast<size_t>(rw) * static_cast<size_t>(cppR),
                     vR.data + static_cast<size_t>(y0 + y) * vR.stride() +
-                        static_cast<size_t>(x0) * cppR,
-                    static_cast<size_t>(rw) * cppR);
-        std::memcpy(subT.buffer->data() + static_cast<size_t>(y) * rw * cppT,
+                        static_cast<size_t>(x0) * static_cast<size_t>(cppR),
+                    static_cast<size_t>(rw) * static_cast<size_t>(cppR));
+        std::memcpy(subT.buffer->data() +
+                        static_cast<size_t>(y) * static_cast<size_t>(rw) * static_cast<size_t>(cppT),
                     vT.data + static_cast<size_t>(y0 + y) * vT.stride() +
-                        static_cast<size_t>(x0) * cppT,
-                    static_cast<size_t>(rw) * cppT);
+                        static_cast<size_t>(x0) * static_cast<size_t>(cppT),
+                    static_cast<size_t>(rw) * static_cast<size_t>(cppT));
     }
     m_ssim = AnalysisEngine::ssim(subR, subT);
     return true;
